+++ OpenAISharp/API/Embeddings.cs
@@ -61,7 +61,7 @@ namespace OpenAISharp.API
             }
             return embeddingsResponse;
         }
-        public static async Task<List<double[]>> Request(string[] input, AvailableModel model = AvailableModel.text_embedding_ada_002)
+        public static async Task<List<float[]>> Request(string[] input, AvailableModel model = AvailableModel.text_embedding_ada_002)
         {
             Embeddings embeddings = new Embeddings() { input = input, SelectedModel =  model};
             EmbeddingsResponse response = await Request(embeddings);
@@ -78,9 +78,9 @@ namespace OpenAISharp.API
             }
             return null;
         }
-        public static async Task<double[]> Request(string input, AvailableModel model = AvailableModel.text_embedding_ada_002)
+        public static async Task<float[]> Request(string input, AvailableModel model = AvailableModel.text_embedding_ada_002)
         {
-            List<double[]> result = await Request(new string[] { input }, model);
+            List<float[]> result = await Request(new string[] { input }, model);
             return result.FirstOrDefault();
         }
     }
